requirements.txt
streamlit>=1.28.0
requests>=2.32.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pandas>=2.0.0
numpy>=1.20.0
openai>=0.28.0